Types

PublicGoodsTournement.AbstractPlayer — Type

AbstractPlayer

An abstract player type for the public goods game. Subtypes of this abstract type must have the following fields

Fields

id: unique player id

trial_start_money::Float64: the amount of money at the beginning of a trial

PublicGoodsTournement.AbstractPublicGoodsGame — Type

AbstractPublicGoodsGame

An abstract public goods game object which expects the following fields.

Fields

total_money::Dict{T, Float64}: the total amount of money for each player

public_goods_multiplier::Float64: the amount by which pooled contributions are multipled

trial_start_money: the amount of money given to an player at the begining of each trial

max_punishment_amount::Float64: the maximum amount that an player can punish on any given trial

PublicGoodsTournement.PublicGoodsGame — Type

PublicGoodsGame{T} <: AbstractPublicGoodsGame

A standard public goods game object with the option for punishment. Punishments are deducted from an player's total money

Fields

total_money::Dict{T, Float64}: the total amount of money for each player

public_goods_multiplier::Float64: the amount by which pooled contributions are multipled

trial_start_money: the amount of money given to an player at the begining of each trial

max_punishment_amount::Float64: the maximum amount that an player can punish on any given trial

Constructors

PublicGoodsGame(; ids, publicgoodsmultiplier, trialstartmoney, maxpunishmentamount )

PublicGoodsGame( totalmoney, publicgoodsmultiplier, trialstartmoney, maxpunishment_amount )

Public Functions

PublicGoodsTournement.contribute — Method

contribute(game_type::Type{<:AbstractPublicGoodsGame}, player::AbstractPlayer)

Contribute to the public good.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::AbstractPlayer: an abstract player type

Returns

contribution::Float64: the amount contributed to the public good

PublicGoodsTournement.contribute — Method

contribute(game_type::Type{<:AbstractPublicGoodsGame}, player::Chris)

Contribute to the public good.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Chris: an abstract player type

Returns

contribution::Float64: the amount contributed to the public good

PublicGoodsTournement.contribute — Method

contribute(game_type::Type{<:AbstractPublicGoodsGame}, player::Dad)

Contribute to the public good.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Dad: an abstract player type

Returns

contribution::Float64: the amount contributed to the public good

PublicGoodsTournement.contribute — Method

contribute(game_type::Type{<:AbstractPublicGoodsGame}, player::DoucheBag)

Contribute to the public good.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::DoucheBag: an abstract player type

Returns

contribution::Float64: the amount contributed to the public good

PublicGoodsTournement.contribute — Method

contribute(game_type::Type{<:AbstractPublicGoodsGame}, player::Mary)

Contribute to the public good.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Mary: an abstract player type

Returns

contribution::Float64: the amount contributed to the public good

PublicGoodsTournement.contribute — Method

contribute(game_type::Type{<:AbstractPublicGoodsGame}, player::Papi)

Contribute to the public good.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Papi: an abstract player type

Returns

contribution::Float64: the amount contributed to the public good

PublicGoodsTournement.punish — Method

punish(game_type::Type{<:AbstractPublicGoodsGame}, player::AbstractPlayer, ids)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::AbstractPlayer: an abstract player type

ids: a collection of player ids

Returns

punishments::Dict{T, Float64}: punishment amount associated with each player: id => punishment

PublicGoodsTournement.punish — Method

punish(game_type::Type{<:AbstractPublicGoodsGame}, player::Chris, ids)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Chris: an abstract player type

ids: a collection of player ids

Returns

punishments::Dict{T, Float64}: punishment amount associated with each player: id => punishment

PublicGoodsTournement.punish — Method

punish(game_type::Type{<:AbstractPublicGoodsGame}, player::Dad, ids)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Dad: an abstract player type

ids: a collection of player ids

Returns

punishments::Dict{T, Float64}: punishment amount associated with each player: id => punishment

PublicGoodsTournement.punish — Method

punish(game_type::Type{<:AbstractPublicGoodsGame}, player::DoucheBag, ids)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::DoucheBag: an abstract player type

ids: a collection of player ids

Returns

punishments::Dict{T, Float64}: punishment amount associated with each player: id => punishment

PublicGoodsTournement.punish — Method

punish(game_type::Type{<:AbstractPublicGoodsGame}, player::Mary, ids)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Mary: an abstract player type

ids: a collection of player ids

Returns

punishments::Dict{T, Float64}: punishment amount associated with each player: id => punishment

PublicGoodsTournement.punish — Method

punish(game_type::Type{<:AbstractPublicGoodsGame}, player::Papi, ids)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Papi: an abstract player type

ids: a collection of player ids

Returns

punishments::Dict{T, Float64}: punishment amount associated with each player: id => punishment

PublicGoodsTournement.setup! — Method

setup!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::AbstractPlayer,
    ids,
    game_config
)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::AbstractPlayer: an abstract player type

ids: a collection of player ids

game_config: keyword arguments of game options

PublicGoodsTournement.setup! — Method

setup!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Chris,
    ids,
    game_config
)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Chris: an abstract player type

ids: a collection of player ids

game_config: keyword arguments of game options

PublicGoodsTournement.setup! — Method

setup!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Dad,
    ids,
    game_config
)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Dad: an abstract player type

ids: a collection of player ids

game_config: keyword arguments of game options

PublicGoodsTournement.setup! — Method

setup!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::DoucheBag,
    ids,
    game_config
)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::DoucheBag: an abstract player type

ids: a collection of player ids

game_config: keyword arguments of game options

PublicGoodsTournement.setup! — Method

setup!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Mary,
    ids,
    game_config
)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Mary: an abstract player type

ids: a collection of player ids

game_config: keyword arguments of game options

PublicGoodsTournement.setup! — Method

setup!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Papi,
    ids,
    game_config
)

Optionally setup player before playing iterated public goods game.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Papi: an abstract player type

ids: a collection of player ids

game_config: keyword arguments of game options

Private Functions

PublicGoodsTournement.observe_contributions! — Method

observe_contributions!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::AbstractPlayer,
    contributions::Dict{T, Float64}
)

Optionally observe each players contribution.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::AbstractPlayer: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_contributions! — Method

observe_contributions!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Chris,
    contributions::Dict{T, Float64}
)

Optionally observe each players contribution.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Chris: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_contributions! — Method

observe_contributions!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Dad,
    contributions::Dict{T, Float64}
)

Optionally observe each players contribution.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Dad: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_contributions! — Method

observe_contributions!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::DoucheBag,
    contributions::Dict{T, Float64}
)

Optionally observe each players contribution.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::DoucheBag: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_contributions! — Method

observe_contributions!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Mary,
    contributions::Dict{T, Float64}
)

Optionally observe each players contribution.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Mary: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_contributions! — Method

observe_contributions!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Papi,
    contributions::Dict{T, Float64}
)

Optionally observe each players contribution.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Papi: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_punishments! — Method

observe_punishments!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::AbstractPlayer,
    punisher_id::T,
    punishment::Dict{T, Float64}
)

Optionally observe the punishments from the punisher.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::AbstractPlayer: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_punishments! — Method

observe_punishments!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Chris,
    punisher_id::T,
    punishment::Dict{T, Float64}
)

Optionally observe the punishments from the punisher.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Chris: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_punishments! — Method

observe_punishments!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Dad,
    punisher_id::T,
    punishment::Dict{T, Float64}
)

Optionally observe the punishments from the punisher.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Dad: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_punishments! — Method

observe_punishments!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::DoucheBag,
    punisher_id::T,
    punishment::Dict{T, Float64}
)

Optionally observe the punishments from the punisher.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::DoucheBag: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_punishments! — Method

observe_punishments!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Mary,
    punisher_id::T,
    punishment::Dict{T, Float64}
)

Optionally observe the punishments from the punisher.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Mary: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.observe_punishments! — Method

observe_punishments!(
    game_type::Type{<:AbstractPublicGoodsGame},
    player::Papi,
    punisher_id::T,
    punishment::Dict{T, Float64}
)

Optionally observe the punishments from the punisher.

Arguments

game_type::Type{<:AbstractPublicGoodsGame}: public goods game type

player::Papi: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing

PublicGoodsTournement.update_money! — Method

update_money!(game::AbstractPublicGoodsGame, contributions)

Update each agents money based on contribution and total public good.

Arguments

player::AbstractPlayer: an abstract player type

Returns

contribution::Float64: the amount contributed to the public good