Types

PublicGoodsTournement.AbstractPlayer — Type

AbstractPlayer

An abstract player type for the public goods game. Subtypes of this abstract type must have the following fields

Fields

id: unique player id

trial_money::Float64: the amount of money at the beginning of a trial

PublicGoodsTournement.AbstractPublicGoodsGame — Type

AbstractPublicGoodsGame

An abstract public goods game object which expects the following fields.

Fields

total_money::Dict{T, Float64}: the total amount of money for each player

public_goods_multiplier::Float64: the amount by which pooled contributions are multipled

trial_start_amount: the amount of money given to an player at the begining of each trial

max_punishment_amount::Float64: the maximum amount that an player can punish on any given trial

PublicGoodsTournement.PublicGoodsGame — Type

PublicGoodsGame{T} <: AbstractPublicGoodsGame

A standard public goods game object with the option for punishment. Punishments are deducted from an player's total money

Fields

total_money::Dict{T, Float64}: the total amount of money for each player

public_goods_multiplier::Float64: the amount by which pooled contributions are multipled

trial_start_amount: the amount of money given to an player at the begining of each trial

max_punishment_amount::Float64: the maximum amount that an player can punish on any given trial

Constructors

PublicGoodsGame(; ids, publicgoodsmultiplier, trialstartamount, maxpunishmentamount )

PublicGoodsGame( totalmoney, publicgoodsmultiplier, trialstartamount, maxpunishment_amount )

Public Functions

PublicGoodsTournement.contribute — Method

contribute(player::AbstractPlayer)

Contribute to the public good.

Arguments

player::AbstractPlayer: an abstract player type

contribute(player::AbstractPlayer)

Returns

contribution::Float64: the amount contributed to the public good

PublicGoodsTournement.punish — Method

punish(player::AbstractPlayer, ids)

Optionally setup player before playing iterated public goods game.

Arguments

player::AbstractPlayer: an abstract player type

ids: a collection of player ids

Returns

punishments::Dict{T, Float64}: punishment amount associated with each player: id => punishment

PublicGoodsTournement.setup! — Method

setup!(player::AbstractPlayer, ids, game_config)

Optionally setup player before playing iterated public goods game.

Arguments

player::AbstractPlayer: an abstract player type

ids: a collection of player ids

game_config: keyword arguments of game options

Private Functions

PublicGoodsTournement.observe_contributions! — Method

observe_contributions!(player::AbstractPlayer, contributions::Dict{T,Float64})

Observe each players contribution.

Arguments

player::AbstractPlayer: an abstract player type

contributions::Dict{T,Float64}: each player's contribution: id => contribution

Returns

nothing